Moved context into its own module.
