+++ compiler/src/lowering/mappers/expressions.rs
@@ -1,6 +1,6 @@
 use crate::lowering::names::tuple_struct_name;
+use fe_analyzer::context::Context;
 use fe_analyzer::namespace::types::Type;
-use fe_analyzer::Context;
 use fe_parser::ast as fe;
 use fe_parser::ast::Kwarg;
 use fe_parser::node::Node;
