+++ compiler/src/yul/mappers/expressions.rs
@@ -7,8 +7,8 @@ use crate::yul::utils::call_arg_value;
 use builtins::{BlockField, ChainField, MsgField, Object, TxField};
 use fe_analyzer::builtins;
 use fe_analyzer::builtins::{ContractTypeMethod, GlobalMethod};
+use fe_analyzer::context::{CallType, Context, Location};
 use fe_analyzer::namespace::types::{Base, FeSized, FixedSize, Type};
-use fe_analyzer::{CallType, Context, Location};
 use fe_common::utils::keccak;
 use fe_parser::ast as fe;
 use fe_parser::node::Node;
